Refactor utils path, add gm bot
