+++ agents/bots/gm/index.ts
@@ -1,0 +1,38 @@
+import { Agent } from "@xmtp/agent-sdk";
+import { getTestUrl, logDetails } from "@xmtp/agent-sdk/debug";
+
+// Load .env file only in local development
+if (process.env.NODE_ENV !== "production") process.loadEnvFile(".env");
+
+const agent = await Agent.createFromEnv({
+  env: process.env.XMTP_ENV as "local" | "dev" | "production",
+});
+
+agent.on("text", async (ctx) => {
+  if (ctx.isDm()) {
+    const messageContent = ctx.message.content;
+    const senderAddress = await ctx.getSenderAddress();
+    console.log(`Received message: ${messageContent} by ${senderAddress}`);
+    await ctx.sendText("gm");
+  }
+});
+
+agent.on("text", async (ctx) => {
+  if (ctx.isGroup() && ctx.message.content.includes("@gm")) {
+    const senderAddress = await ctx.getSenderAddress();
+    console.log(
+      `Received message in group: ${ctx.message.content} by ${senderAddress}`,
+    );
+    await ctx.sendText("gm");
+  }
+});
+
+agent.on("start", () => {
+  console.log(`Waiting for messages...`);
+  console.log(`Address: ${agent.address}`);
+  console.log(agent.client.options);
+  console.log(`🔗${getTestUrl(agent.client)}`);
+  void logDetails(agent.client);
+});
+
+await agent.start();
