+++ agents/bots/key-check/handlers/ux.ts
@@ -5,7 +5,7 @@ import {
   type RemoteAttachment,
 } from "@xmtp/content-type-remote-attachment";
 import { ContentTypeWalletSendCalls } from "@xmtp/content-type-wallet-send-calls";
-import { USDCHandler } from "../../utils/usdc";
+import { USDCHandler } from "../../../utils/usdc";
 
 export class UxHandlers {
   private usdcHandler: USDCHandler;
