+++ agents/utils/inline-actions/types/ActionsContent.ts
