+++ agents/utils/inline-actions/types/IntentContent.ts
